docs: renumber Pydantic guide to 11
