+++ docs/03_guides/11_pydantic.mdx
@@ -1,6 +1,6 @@
 ---
 id: input-validation
-title: Validate Actor input with Pydantic
+title: Input validation with Pydantic
 description: Parse, validate, and type your Actor's input with Pydantic models instead of reaching into a raw dictionary.
 ---
 
@@ -8,10 +8,10 @@ import CodeBlock from '@theme/CodeBlock';
 import RunnableCodeBlock from '@site/src/components/RunnableCodeBlock';
 import ApiLink from '@theme/ApiLink';
 
-import RawInputExample from '!!raw-loader!roa-loader!./code/10_raw_input.py';
-import PydanticExample from '!!raw-loader!roa-loader!./code/10_pydantic.py';
-import HttpUrlExample from '!!raw-loader!./code/10_http_url.py';
-import ModelValidatorExample from '!!raw-loader!./code/10_model_validator.py';
+import RawInputExample from '!!raw-loader!roa-loader!./code/11_raw_input.py';
+import PydanticExample from '!!raw-loader!roa-loader!./code/11_pydantic.py';
+import HttpUrlExample from '!!raw-loader!./code/11_http_url.py';
+import ModelValidatorExample from '!!raw-loader!./code/11_model_validator.py';
 
 In this guide, you'll learn how to validate your Apify Actor's input with [Pydantic](https://docs.pydantic.dev/), so that your code works with a typed, guaranteed-valid object instead of a raw dictionary.
