+++ docs/03_guides/code/11_raw_input.py
@@ -6,7 +6,7 @@ from apify import Actor
 async def main() -> None:
     # Enter the context of the Actor.
     async with Actor:
-        # Read the input and reach into the raw dictionary for each value.
+        # Read the input and reach into the raw dict.
         actor_input = await Actor.get_input() or {}
         search_terms = actor_input.get('searchTerms', [])
         max_results = actor_input.get('maxResults', 10)
